+++ client/src/App.js
@@ -8,7 +8,7 @@ import Login from "../src/pages/Login,SignUp page/Login";
 import SignUp from "../src/pages/Login,SignUp page/SignUp";
 import Mypage from "../src/pages/Mypage";
 import Detail from "../src/pages/Detail";
-import NewPost from "../src/pages/PostPage/NewPost";
+import Post from "../src/pages/PostPage/Post";
 import Result from "../src/pages/Result";
 import Footer from "./components/Footer";
 import "./App.css";
@@ -25,7 +25,7 @@ const App = () => {
                 <Route path="/signup" element={<SignUp />}></Route>
                 <Route path="/mypage" element={<Mypage />}></Route>
                 <Route path="/detail/:contentId" element={<Detail />}></Route>
-                <Route path="/post" element={<NewPost />}></Route>
+                <Route path="/post" element={<Post />}></Route>
                 <Route path="/result" element={<Result />}></Route>
             </Routes>
         </BrowserRouter>
